+++ Tests/BugsnagBaseUnitTest.m
@@ -26,7 +26,7 @@
  *
  * We take the former approach.
  *
- * @param willNotify Whether the notifier should actually send the event to the server
+ * @param willNotify Whether the Notifier should actually send the event to the server
  * @param willPersistUser Whether any user information should be persisted
  */